floater/requirements.txt
paho-mqtt
python-dotenv
influxdb-client
scapy